+++ src/shock95x/auctionhouse/economy/EconomyProvider.php
@@ -7,9 +7,9 @@ use pocketmine\player\Player;
 
 interface EconomyProvider {
 
-	public function addMoney(string|Player $player, float $amount, callable $callback): void;
+	public function addMoney(string|Player $player, int $amount, callable $callback): void;
 
-	public function subtractMoney(string|Player $player, float $amount, callable $callback): void;
+	public function subtractMoney(string|Player $player, int $amount, callable $callback): void;
 
 	/**
 	 * Get currency symbol of economy provider
